<fix> 커뮤니티 모듈 제거
