--- src/components/community/post/Create.jsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const Create = () => {
-  return <div>Create</div>;
-};
-
-export default Create;
