--- src/components/community/post/ImageView.jsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const ImageView = () => {
-  return <div>ImageView</div>;
-};
-
-export default ImageView;
