--- src/page/community/Detail.jsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const Detail = () => {
-  return <div>Detail</div>;
-};
-
-export default Detail;
