--- src/page/community/Main.jsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const Main = () => {
-  return <div>Main</div>;
-};
-
-export default Main;
