--- src/page/community/Posting.jsx
@@ -1,17 +1,0 @@
-import React from 'react';
-import Footer from '../../components/elements/GlobalFooter';
-import Header from '../../components/elements/GlobalHeader';
-import Layout from '../../components/elements/GlobalLayout';
-import Create from '../../components/market/Create';
-
-const Posting = () => {
-  return (
-    <Layout>
-      <Header />
-      <Create />
-      <Footer />
-    </Layout>
-  );
-};
-
-export default Posting;
