--- src/page/community/PostingUpdate.jsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const PostingUpdate = () => {
-  return <div>PostingUpdate</div>;
-};
-
-export default PostingUpdate;
